+++ URAN-2017/ClassBAAK12NoTail.cs
@@ -1762,7 +1762,7 @@ namespace URAN_2017
         }
 
 
-        private void BDReadСобытие(string nameFile, string nameBAAK, string time, string nameRan, int[] Amp, string nameklaster, int[] Nnut, int[] Nl, Double[] sig, bool test)
+        private void BDReadСобытие(string nameFile, string nameBAAK, string time, string nameRan, int[] Amp, string nameklaster, int[] Nnut, double[] Nl, Double[] sig, bool test)
         {
             if (FlagSaveBD)
             {
